refactor: consolidate file handling & reporting
